+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Automatic handling of pull request events to use correct commit SHA and branch name
+- Uses PR head commit SHA instead of merge commit SHA
+- Uses PR source branch name instead of target branch
+
 ## [1.6.0] - 2024-03-22
 
 ### Changed